pyproject.toml
[project]
name = "lawcite"
version = "0.1.3"
description = "Tools for working with Danish laws in LaTeX"
authors = [{ name = "evid 210122431+evidlabel@users.noreply.github.com" }]
requires-python = "~=3.12.0"
readme = "README.md"
license = "MIT"
dependencies = [
    "bibtexparser>=1.4.1,<2",
    "requests>=2.31.0,<3",
    "pypdf>=5.8.0,<6",
    "unidecode>=1.3.8,<2",
    "pyyaml>=6.0.1,<7",
    "treeparse",
    "pydantic>=2.0.0,<3",
]

[project.scripts]
lawcite = "lawcite.cli.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0,<9",
    "pytest-cov>=4.1.0,<5",
    "black>=24.0.0,<25",
    "flake8>=7.0.0,<8",
    "mkdocs>=1.5.3,<2",
    "mkdocs-material>=9.5.0,<10",
]

[tool.hatch.build.targets.sdist]
include = ["src/lawcite"]

[tool.hatch.build.targets.wheel]
include = ["src/lawcite"]

[tool.hatch.build.targets.wheel.sources]
"src/lawcite" = "lawcite"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
